PSR-1984: create utils for asset classes and refactor CYA controllers
(land/property, loans, shares, bonds, otherassets)
