+++ app/views/SummaryView.scala.html
@@ -1,0 +1,65 @@
+@*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import viewmodels.models.CheckYourAnswersSummaryViewModel
+@import viewmodels.models.SummaryPageEntry.Section
+
+@this(
+        layout: templates.Layout,
+        formHelper: FormWithCSRF,
+        govukSummaryList: GovukSummaryList,
+        govukInsetText: GovukInsetText,
+        govukPagination: GovukPagination,
+        govukButton: GovukButton,
+)
+
+    @(viewmodel: FormPageViewModel[List[SummaryPageEntry]], schemeName: String)(implicit request: Request[?], messages: Messages)
+
+    @layout(pageTitle = titleNoForm( if (viewmodel.mode == ViewOnlyMode && viewmodel.optViewOnlyDetails.nonEmpty) viewmodel.optViewOnlyDetails.get.title.toMessage else viewmodel.title.toMessage),showBackLink = viewmodel.showBackLink) {
+
+        <span class="govuk-caption-xl">@schemeName</span>
+
+        @if(viewmodel.mode == NormalMode || viewmodel.mode == CheckMode) {
+            <h1 class="govuk-heading-xl">@renderMessage(viewmodel.heading)</h1>
+        }
+
+        @for(vm <- viewmodel.page) {
+            @vm match {
+                case Section(vmSection) => {
+                    <h1 class="govuk-heading-m">@renderMessage(vmSection.heading)</h1>
+                    @vmSection.sections.map { section =>
+                        @section.heading.map(renderMessage)
+                        @govukSummaryList(
+                            SummaryListViewModel(
+                                section.rows.map { row =>
+                                    SummaryListRowViewModel(
+                                        key = if(row.oneHalfWidth) KeyViewModel(row.key.toMessage).withOneHalfWidth() else KeyViewModel(row.key.toMessage),
+                                        value = ValueViewModel(renderMessage(row.value)),
+                                        actions = Seq()
+                                    )
+                                }
+                            ).withMarginBottom(vmSection.marginBottom)
+                        )
+                    }
+
+
+                    @vmSection.inset.map { inset =>
+                        @govukInsetText(InsetTextViewModel(renderMessage(inset)))
+                    }
+                }
+            }
+        }
+    }
