+++ conf/messages.en
@@ -1242,6 +1242,8 @@ nonsipp.tasklist.declaration.complete = Read declaration and submit
 nonsipp.tasklist.declaration.view = View submission confirmation
 nonsipp.tasklist.declaration.cip.complete = Declaration
 nonsipp.tasklist.declaration.saveandreturn = Save and return to {0} overview
+nonsipp.summary.title = Pension Scheme Return summary
+nonsipp.summary.heading = Pension Scheme Return summary
 
 quotedSharesManagedFundsHeld.title = Did the scheme hold any quoted shares or third party managed funds on {0}?
 quotedSharesManagedFundsHeld.heading = Did {0} hold any quoted shares or third party managed funds on {1}?
